+++ NEWS.md
@@ -1,3 +1,9 @@
+# metapip (development version)
+
+# metapip 0.0.0.9010
+
+* Add a new `package_branches()` function to show package versions in branches.
+
 # metapip 0.0.0.9008
 
 * Unload packages before installing (#3)